refactor: split skills into discovery and index-store
